+++ samples/node/app.js
@@ -19,10 +19,10 @@ var residueParams = {
         }
     ],
     application_id: "com.muflihun.securitybox",
-    client_id: "muflihun00102030",
     rsa_key_size: 2048,
     utc_time: true,
     time_offset: 0,
+    client_id: "muflihun00102030",
     client_private_key: "/Users/majid.khan/Projects/residue/samples/clients/netcat/client-256-private.pem",
     server_public_key: "/Users/majid.khan/Projects/residue/samples/clients/netcat/server-1024-public.pem"
 };